# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Social Reaction Analysis GR",
    "description": "Real-time detection and visualisation of social reactions in Greece.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/events": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "List Events",
        "operationId": "list_events_events_get",
        "parameters": [
          {
            "name": "action_form",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action Form"
            }
          },
          {
            "name": "thematic_field",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Thematic Field"
            }
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Channel"
            }
          },
          {
            "name": "intensity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Intensity"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO 8601 date",
              "title": "Date From"
            },
            "description": "ISO 8601 date"
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO 8601 date",
              "title": "Date To"
            },
            "description": "ISO 8601 date"
          },
          {
            "name": "bbox",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "west,south,east,north",
              "title": "Bbox"
            },
            "description": "west,south,east,north"
          },
          {
            "name": "temporal_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "upcoming",
                    "today",
                    "past"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Temporal Status"
            }
          },
          {
            "name": "is_national",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is National"
            }
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "recent",
                "event_time"
              ],
              "type": "string",
              "default": "recent",
              "title": "Order By"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventSummary"
                  },
                  "title": "Response List Events Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/events/geojson": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Events Geojson",
        "operationId": "events_geojson_events_geojson_get",
        "parameters": [
          {
            "name": "action_form",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action Form"
            }
          },
          {
            "name": "thematic_field",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Thematic Field"
            }
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Channel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoJSONFeatureCollection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/events/{event_id}": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get Event",
        "operationId": "get_event_events__event_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ArticleSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "source_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Type"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "ArticleSummary"
      },
      "EventDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "action_forms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Action Forms"
          },
          "thematic_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Thematic Fields"
          },
          "channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "intensity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intensity"
          },
          "summary_el": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary El"
          },
          "summary_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary En"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lon": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lon"
          },
          "article_count": {
            "type": "integer",
            "title": "Article Count"
          },
          "source_count": {
            "type": "integer",
            "title": "Source Count"
          },
          "first_seen": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Seen"
          },
          "last_seen": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "event_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Time"
          },
          "temporal_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temporal Status"
          },
          "is_national": {
            "type": "boolean",
            "title": "Is National",
            "default": false
          },
          "announced_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Announced By"
          },
          "participating_unions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Participating Unions",
            "default": []
          },
          "classification_confidence": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classification Confidence"
          },
          "articles": {
            "items": {
              "$ref": "#/components/schemas/ArticleSummary"
            },
            "type": "array",
            "title": "Articles",
            "default": []
          },
          "reactions": {
            "items": {
              "$ref": "#/components/schemas/ReactionSummary"
            },
            "type": "array",
            "title": "Reactions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "action_forms",
          "thematic_fields",
          "article_count",
          "source_count",
          "status"
        ],
        "title": "EventDetail"
      },
      "EventSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "action_forms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Action Forms"
          },
          "thematic_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Thematic Fields"
          },
          "channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "intensity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intensity"
          },
          "summary_el": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary El"
          },
          "summary_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary En"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lon": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lon"
          },
          "article_count": {
            "type": "integer",
            "title": "Article Count"
          },
          "source_count": {
            "type": "integer",
            "title": "Source Count"
          },
          "first_seen": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Seen"
          },
          "last_seen": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "event_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Time"
          },
          "temporal_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temporal Status"
          },
          "is_national": {
            "type": "boolean",
            "title": "Is National",
            "default": false
          },
          "announced_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Announced By"
          },
          "participating_unions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Participating Unions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "action_forms",
          "thematic_fields",
          "article_count",
          "source_count",
          "status"
        ],
        "title": "EventSummary"
      },
      "GeoJSONFeature": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "Feature"
          },
          "geometry": {
            "$ref": "#/components/schemas/GeoJSONGeometry"
          },
          "properties": {
            "$ref": "#/components/schemas/GeoJSONProperties"
          }
        },
        "type": "object",
        "required": [
          "geometry",
          "properties"
        ],
        "title": "GeoJSONFeature"
      },
      "GeoJSONFeatureCollection": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "FeatureCollection"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/GeoJSONFeature"
            },
            "type": "array",
            "title": "Features"
          }
        },
        "type": "object",
        "required": [
          "features"
        ],
        "title": "GeoJSONFeatureCollection"
      },
      "GeoJSONGeometry": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "Point"
          },
          "coordinates": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Coordinates"
          }
        },
        "type": "object",
        "required": [
          "coordinates"
        ],
        "title": "GeoJSONGeometry"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}